Backport the mcmod.info ResourceLoading fix to FG1 projects
