+++ src/minecraft/kotlin/xyz/wagyourtail/unimined/internal/minecraft/patch/forge/fg1/FG1MinecraftTransformer.kt
@@ -13,12 +13,15 @@ import xyz.wagyourtail.unimined.api.minecraft.MinecraftJar
 import xyz.wagyourtail.unimined.internal.minecraft.patch.forge.ForgeLikeMinecraftTransformer
 import xyz.wagyourtail.unimined.internal.minecraft.patch.jarmod.JarModAgentMinecraftTransformer
 import xyz.wagyourtail.unimined.internal.minecraft.resolver.Library
+import xyz.wagyourtail.unimined.internal.minecraft.transform.fixes.FixFG1ResourceLoading
+import xyz.wagyourtail.unimined.internal.minecraft.transform.fixes.FixFG1ResourceLoading.fixResourceLoading
 import xyz.wagyourtail.unimined.internal.minecraft.transform.merge.ClassMerger
 import xyz.wagyourtail.unimined.mapping.Namespace
 import xyz.wagyourtail.unimined.util.*
 import xyz.wagyourtail.unimined.util.deleteRecursively
 import java.io.File
 import java.io.InputStream
+import java.nio.file.FileSystem
 import java.nio.file.Files
 import java.nio.file.StandardCopyOption
 import java.nio.file.StandardOpenOption
@@ -57,6 +60,10 @@ open class FG1MinecraftTransformer(project: Project, val parent: ForgeLikeMinecr
             getDynLibs(it)
         }
     }
+
+    override val transform = (listOf<(FileSystem) -> Unit>(
+        FixFG1ResourceLoading::fixResourceLoading
+    ) + super.transform).toMutableList()
 
     override fun apply() {
         // get and add forge-src to mappings
@@ -227,6 +234,10 @@ open class FG1MinecraftTransformer(project: Project, val parent: ForgeLikeMinecr
     override fun applyClientRunTransform(config: RunConfig) {
         super.applyClientRunTransform(config)
 
+        config.properties["source_roots"] = {
+            parent.groups
+        }
+
         // resolve dyn libs
         val dynLibFolder = config.workingDir.resolve("lib")
         dynLibFolder.mkdirs()
@@ -238,10 +249,15 @@ open class FG1MinecraftTransformer(project: Project, val parent: ForgeLikeMinecr
 
         config.jvmArgs("-Dminecraft.applet.TargetDirectory=${config.workingDir.absolutePath}")
         if (parent.mainClass != null) config.mainClass.set(parent.mainClass!!)
+        config.environment["MOD_CLASSES"] = "\${source_roots}"
     }
 
     override fun applyServerRunTransform(config: RunConfig) {
         super.applyServerRunTransform(config)
+
+        config.properties["source_roots"] = {
+            parent.groups
+        }
 
         // resolve dyn libs
         val dynLibFolder = config.workingDir.resolve("lib")
@@ -253,6 +269,7 @@ open class FG1MinecraftTransformer(project: Project, val parent: ForgeLikeMinecr
         }
 
         if (parent.mainClass != null) config.mainClass.set(parent.mainClass!!)
+        config.environment["MOD_CLASSES"] = "\${source_roots}"
     }
 
     override fun afterRemap(baseMinecraft: MinecraftJar): MinecraftJar {
